front-end files partly separated by folders
