+++ ShareCar.Client/src/components/Driver/Driver.jsx
@@ -2,17 +2,19 @@
 import React, { Component } from "react";
 import { Route, Link } from "react-router-dom";
 import { Router, Switch } from "react-router";
-import UserService from "../services/userService";
-import AuthenticationService from "../services/authenticationService";
-import history from "../helpers/history";
+import UserService from "../../services/userService";
+import AuthenticationService from "../../services/authenticationService";
+import history from "../../helpers/history";
 import { ViewRideRequests } from "./ViewRideRequests";
-import { NavBar } from "./NavBar";
+
 export default class Driver extends Component<{}, MyProfileState> {
   userService = new UserService();
   state: MyProfileState = { loading: true, user: null };
 
   componentDidMount() {
     this.userService.getLoggedInUser(this.updateLoggedInUser);
+    console.log(this.props.isDriver);
+    this.props.isDriver == null ? history.push("/") : null;
   }
 
   updateLoggedInUser = (user: UserProfileData) => {
@@ -28,7 +30,8 @@ export default class Driver extends Component<{}, MyProfileState> {
       <p>Failed</p>
     ) : (
       <div className="role-container">
-        <ViewRideRequests driver={false} />
+        <h1 className="role-container">Driver Page</h1>
+        <ViewRideRequests driver={true} />
       </div>
     );
     return <div>{content}</div>;
