+++ ShareCar.Client/src/components/Passenger/Passenger.jsx
@@ -1,21 +1,18 @@
 // @flow
 import React, { Component } from "react";
-import { Route, Link } from "react-router-dom";
-import { Router, Switch } from "react-router";
-import UserService from "../services/userService";
-import AuthenticationService from "../services/authenticationService";
-import history from "../helpers/history";
-import { ViewRideRequests } from "./ViewRideRequests";
-import { NavBar } from "./NavBar";
-
+//import { Route, Link } from "react-router-dom";
+//import { Router, Switch } from "react-router";
+import UserService from "../../services/userService";
+//import AuthenticationService from "../services/authenticationService";
+//import history from "../helpers/history";
+import { ViewRideRequests } from "../Driver/ViewRideRequests";
+//import { NavBar } from "./NavBar";
 export default class Driver extends Component<{}, MyProfileState> {
   userService = new UserService();
   state: MyProfileState = { loading: true, user: null };
 
   componentDidMount() {
     this.userService.getLoggedInUser(this.updateLoggedInUser);
-    console.log(this.props.isDriver);
-    this.props.isDriver == null ? history.push("/") : null;
   }
 
   updateLoggedInUser = (user: UserProfileData) => {
@@ -31,8 +28,7 @@ export default class Driver extends Component<{}, MyProfileState> {
       <p>Failed</p>
     ) : (
       <div className="role-container">
-        <h1 className="role-container">Driver Page</h1>
-        <ViewRideRequests driver={true} />
+        <ViewRideRequests driver={false} />
       </div>
     );
     return <div>{content}</div>;
