+++ ShareCar.Client/src/components/Passenger/PassengerRideRequestsList.js
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { Status } from "./status";
-import "../styles/riderequests.css";
+import "../../styles/riderequests.css";
 import Moment from "react-moment";
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
@@ -10,9 +10,9 @@ import TableRow from '@material-ui/core/TableRow';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import { withStyles } from '@material-ui/core/styles';
-import "../styles/genericStyles.css";
-import { Typography } from "../../node_modules/@material-ui/core";
-import MapComponent from "./MapComponent";
+import "../../styles/genericStyles.css";
+import { Typography } from "../../../node_modules/@material-ui/core";
+import MapComponent from "../Maps/MapComponent";
 import Button from "@material-ui/core/Button";
 
 
@@ -51,9 +51,7 @@ state={
                           Show on map
                         </Button>
                         
-                             {   /*<Typography component="p">
-                                    Address: {      }
-                    </Typography>*/}
+
                             </CardContent>
                         </tr>
                     )}
