+++ ShareCar.Client/src/components/WinnersList.jsx
@@ -2,12 +2,12 @@ import * as React from "react";
 import axios from "axios";
 import api from "../helpers/axiosHelper";
 import "../styles/driversRidesList.css";
-import { Route, Link } from "react-router-dom";
-import NewRideForm from "./NewRideForm";
-import PassengersList from "./PassengersList";
+//import { Route, Link } from "react-router-dom";
+//import NewRideForm from "./NewRideForm";
+//import PassengersList from "./PassengersList";
 import "../styles/genericStyles.css";
 import "../styles/driversRidesList.css";
-import { ViewRideRequests } from "../components/ViewRideRequests";
+//import { ViewRideRequests } from "../components/ViewRideRequests";
 import Card from "@material-ui/core/Card";
 import Typography from "@material-ui/core/Typography";
 import CardActions from "@material-ui/core/CardActions";
