+++ ShareCar.Client/src/index.js
@@ -7,22 +7,15 @@ import Layout from "./components/common/Layout";
 import Login from "./components/Login";
 import MyProfile from "./components/MyProfile";
 import RoleSelection from "./components/RoleSelection";
-import Driver from "./components/Driver";
-import Passenger from "./components/Passenger";
+//import Driver from "./Driver/components/Driver";
+import Passenger from "./components/Passenger/Passenger";
 import UserProfile from "./components/UserProfile";
-import MapComponent from "./components/MapComponent";
 import Rides from "./components/Rides";
-import NewRideForm from "./components/NewRideForm";
-import RideRequestForm from "./components/RideRequestForm";
 import WinnerBoard from "./components/WinnerBoard";
-import test from "./components/TestMap";
-import DriverMap from "./components/DriverMap";
-import PassengerMap from "./components/PassengerMap";
-import Map from "./components/Map";
+import Map from "./components/Maps/Map";
 import SnackBars from "./components/common/Snackbars";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./index.css";
-import RidesScheduler from "./components/RidesScheduler";
 
 const options = {
   position: "bottom center",
